Add closing capability and improved "already open checks"
Tabs (i.e. open Profiles) can now be closed. When closed, the underlying
monitoring will be closed as well, and the resources (i.e.
ProfileContext etc.) will be released.

Also, for large profiles it was possible to open the same file twice
before the first one was fully processed.

However as a side effect, currently when a diff has been opened for 2
profiles A and B, and the "non-diff" views for A or B are closed, these
can not be reopened while the diff is open. Close the diff first before
trying to reopen the files for A or B.