Reenable VM item in monitor menu when profile closed
Part of tab-closing logic for live profiles.
It's still possible to open the logfile for a VM already being monitored
through the file menu though. I'll probably have to revisit the "source"
logic for cleanup, so that there is a single place which keeps track of
what has been opened or what is being opened.