Improved "check-whether-already-opened" logic
The logic for checking whether a file was already open in a
ProfileContext was incomplete. Symlinks weren't being resolved, which
could lead to duplicate profiles based on the same underlying file.

This is now working properly. A refactor is needed however to clean the
logic up. Specifically, the ProfileContext should probably be registered
right away when an attempt for profiling is made, and unregistered if
the attempt fails. Also, the logsource should provide its own id.